requirements.txt
# Базовые зависимости (без сборки C-расширений — ставится на Windows без Visual C++)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
openai>=1.12.0
python-multipart>=0.0.6
pydantic-settings>=2.1.0
# Извлечение текста из документов
pypdf>=4.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
